+++ linkerd/protocol/http/src/test/scala/com/twitter/finagle/buoyant/linkerd/HttpTraceInitializerTest.scala
@@ -1,5 +1,6 @@
 package com.twitter.finagle.buoyant.linkerd
 
+import com.twitter.finagle.Stack
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.tracing.{NullTracer, Trace}
 import com.twitter.finagle.{Service, ServiceFactory, param}
@@ -13,11 +14,12 @@ class HttpTraceInitializerTest extends FunSuite with Awaits {
   def tracingService(): Service[Request, Response] = {
     val svc = Service.mk[Request, Response] { req =>
       val rsp = Response()
-      Headers.Ctx.set(rsp.headerMap, Trace.id)
+      Headers.Ctx.Trace.set(rsp.headerMap, Trace.id)
       Future.value(rsp)
     }
-    val tracer = param.Tracer(NullTracer)
-    val factory = HttpTraceInitializer.server.make(tracer, ServiceFactory.const(svc))
+    val factory = HttpTraceInitializer.serverModule
+      .toStack(Stack.Leaf(Stack.Role("ep"), ServiceFactory.const(svc)))
+      .make(Stack.Params.empty + param.Tracer(NullTracer))
     Await.result(factory())
   }
 
@@ -26,7 +28,7 @@ class HttpTraceInitializerTest extends FunSuite with Awaits {
     val req = Request()
 
     val rsp = await(service(req))
-    assert(Headers.Ctx.get(rsp.headerMap).exists(_.sampled.isEmpty))
+    assert(Headers.Ctx.Trace.get(rsp.headerMap).exists(_.sampled.isEmpty))
   }
 
   test("does not trace when sample header is set to 0") {
@@ -35,7 +37,7 @@ class HttpTraceInitializerTest extends FunSuite with Awaits {
     req.headerMap(Headers.Sample.Key) = "0"
 
     val rsp = await(service(req))
-    assert(Headers.Ctx.get(rsp.headerMap).exists(_.sampled.contains(false)))
+    assert(Headers.Ctx.Trace.get(rsp.headerMap).exists(_.sampled.contains(false)))
   }
 
   test("traces when sample header is set to 1") {
@@ -44,19 +46,19 @@ class HttpTraceInitializerTest extends FunSuite with Awaits {
     req.headerMap(Headers.Sample.Key) = "1"
 
     val rsp = await(service(req))
-    assert(Headers.Ctx.get(rsp.headerMap).exists(_.sampled.contains(true)))
+    assert(Headers.Ctx.Trace.get(rsp.headerMap).exists(_.sampled.contains(true)))
   }
 
   test("parent/child requests are assigned linked request ids") {
     val service = tracingService()
     val req1 = Request()
     val rsp1 = await(service(req1))
-    val reqId1 = Headers.Ctx.get(rsp1.headerMap).get
+    val reqId1 = Headers.Ctx.Trace.get(rsp1.headerMap).get
 
     val req2 = Request()
-    req2.headerMap(Headers.Ctx.Key) = rsp1.headerMap(Headers.Ctx.Key)
+    req2.headerMap(Headers.Ctx.Trace.Key) = rsp1.headerMap(Headers.Ctx.Trace.Key)
     val rsp2 = await(service(req2))
-    val reqId2 = Headers.Ctx.get(rsp2.headerMap).get
+    val reqId2 = Headers.Ctx.Trace.get(rsp2.headerMap).get
 
     assert(reqId1.spanId == reqId1.parentId)
     assert(reqId2.spanId != reqId2.parentId)
@@ -69,9 +71,9 @@ class HttpTraceInitializerTest extends FunSuite with Awaits {
     val req = Request()
 
     val rsp1 = await(service(req))
-    val reqId1 = Headers.Ctx.get(rsp1.headerMap).get
+    val reqId1 = Headers.Ctx.Trace.get(rsp1.headerMap).get
     val rsp2 = await(service(req))
-    val reqId2 = Headers.Ctx.get(rsp2.headerMap).get
+    val reqId2 = Headers.Ctx.Trace.get(rsp2.headerMap).get
 
     assert(reqId1.traceId != reqId2.traceId)
     assert(reqId1.spanId != reqId2.spanId)