+++ linkerd/protocol/http/src/test/scala/io/buoyant/linkerd/protocol/http/ErrorResponderTest.scala
@@ -2,11 +2,12 @@ package io.buoyant.linkerd.protocol.http
 
 import com.twitter.finagle.{Service, ServiceFactory, Stack}
 import com.twitter.finagle.buoyant.linkerd.Headers
-import com.twitter.util.Future
 import com.twitter.finagle.http.{Request, Response, Status}
-import java.net.URLEncoder
+import com.twitter.io.Charsets
+import com.twitter.util.Future
 import io.buoyant.router.RoutingFactory
 import io.buoyant.test.Awaits
+import java.net.URLEncoder
 import org.scalatest.FunSuite
 
 class ErrorResponderTest extends FunSuite with Awaits {
@@ -14,8 +15,8 @@ class ErrorResponderTest extends FunSuite with Awaits {
   val svc = Service.mk[Request, Response] { _ =>
     Future.exception(RoutingFactory.UnknownDst(Request(), new Exception(s"foo\nbar")))
   }
-  val stk = ErrorResponder.toStack(
-    Stack.Leaf(ErrorResponder.role, ServiceFactory.const(svc))
+  val stk = ErrorResponder.module.toStack(
+    Stack.Leaf(Stack.Role("endpoint"), ServiceFactory.const(svc))
   )
   val service = await(stk.make(Stack.Params.empty)())
 
@@ -26,8 +27,8 @@ class ErrorResponderTest extends FunSuite with Awaits {
 
   test("correctly encodes error headers") {
     val rsp = await(service(Request()))
-    val headerErr = rsp.headerMap(Headers.Err)
+    val headerErr = rsp.headerMap(Headers.Err.Key)
     assert(!headerErr.contains("\n"))
-    assert(headerErr.contains(URLEncoder.encode("\n", "ISO-8859-1")))
+    assert(headerErr.contains(URLEncoder.encode("\n", Charsets.Iso8859_1.toString)))
   }
 }